ANDROID: drm/xe/pxp_multi: Add multi-session management
Now that we have clients, we can allow them to reserve and release
PXP sessions via a new action in the PXP OPS ioctl. A tag to identify
the session instance is returned when the session is reserved.
Reservation is only available while the device is awake and the default
session is automatically started when reservation occurs.

All sessions are automatically released during a PXP termination.

v2: better locking code, avoid atomic operations

Bug: 415852458
Bug: 470437544
Bug: 472380844
Bug: 478238685
Test: DUT booted to UI and Test PAVP IGT tests.
Upstream-Task: 454463101
Upstream-Task: 478238685

Change-Id: If7be671713ee12ca0bede8eb5b897c41ac22d711
Signed-off-by: Daniele Ceraolo Spurio <daniele.ceraolospurio@intel.com>
Signed-off-by: George D Sworo <george.d.sworo@intel.corp-partner.google.com>