ANDROID: x86/pkvm: Use hypercalls to access MMIO
Emulated MMIO cannot work for protected VMs on x86 in the usual way,
since in order to emulate guest's access to an MMIO memory address, the
host needs to read and decode the trapped instruction from the guest
memory (to find out what kind of memory access it is, i.e. is it read or
write and how many bytes), while the host has no access to protected VM
memory.

So overwrite the MMIO PV ops with pkvm's virt MMIO operations which will
use special hypercalls to access MMIO. pKVM will forward handling those
hypercalls to the host, to let the host emulate MMIO access without the
need to decode instructions.

[ Dmytro: fix error handling + extend commit message ]

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: None
Change-Id: I0362f49e27f2b2967f1e99fc414be2fe6cbda0d8