ANDROID: pKVM: x86: Enforce secure pvmfw bootstrap
When a pVM starts with pvmfw, enforce starting it with a known secure
state of its control regs (CR0/CR4/EFER/RFLAGS), segment regs and
GDT/IDT/LDT/TSS.

Most importantly, enforce starting it in 32-bit protected mode with
paging disabled (with the assumption that it is only for bootstrap, and
pvmfw itself will enable paging and switch to 64-bit long mode as early
as possible), similarly to TDX.

The reason for doing that is that 64-bit mode requires paging enabled,
and we cannot trust the initial page tables in the guest memory prepared
by the host VMM (e.g. we need to be sure that the pvmfw entry address in
RIP is not mapped to some other address), so pKVM itself would need to
enforce the initial page tables somewhere in the guest memory, which
would complicate memory management and require a memory layout contract
between pKVM, the VMM and pvmfw.

Of course the initial segment descriptors in the GDT prepared by the
host cannot be trusted either, but we don't need to enforce the initial
GDT in memory, we can just enforce the cached segment descriptors via
VMCS. The CPU will use those cached descriptors instead of GDT until the
guest reloads a segment register (so another assumption is that pvmfw
will avoid reloading segment registers before it sets up its own GDT).
Set the initial GDTR to not point to a valid GDT, so that if the GDT in
memory is still accidentally used too early, it generates a #GP.

While we're at it, since we cannot trust the initial IDT either, set the
initial IDTR to an invalid value as well, so any early exception (#GP in
the above case or any else) will result in a triple fault, rather than
executing an untrusted exception handler prepared by the host.

Since paging is disabled, don't care to set CR3.

[ Grzegorz: squash with:
  ANDROID: pkvm: x86: Enforce initial CR0/CR4/EFER/RFLAGS for pvmfw
  https://android-review.googlesource.com/c/kernel/common/+/3852467
  and move some commit log details from the squashed patch ]

Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Grzegorz Jaszczyk <jaszczyk@google.com>

Bug: None
Change-Id: I95a3bed93bb40f6b3bf497207cdc6b5edc8678ff