ANDROID: pKVM: x86: Implement copying pvmfw into pVM memory
While most of the initial contents of pVM memory is prepared by the
host, pvmfw is an exception: it must be loaded into pVM memory by the
pKVM hypervisor, not by the host which is untrusted and has no access
to the pvmfw memory.

So implement loading pvmfw into pVM memory, similarly to pKVM-ARM: when
donating a page which overlaps with the pvmfw address region in the
guest memory (specified previously by the host), copy the corresponding
portion of the pvmfw image (resident in the hypervisor's reserved
memory) into this newly donated guest page.

This ensures the pVM starts with the correct firmware code in memory.

[ Grzegorz: align to 6.18: s/kvm_protected_vm/kvm_pkvm_vm/; update the
  commit log ]

Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Grzegorz Jaszczyk <jaszczyk@google.com>

Bug: None
Change-Id: I882e4a3577260d2970ed9b318f43f65dc0c99b8b