ANDROID: pKVM: x86: Add spinlock for the pKVM hypervisor
As the pKVM hypervisor will run in the root mode, using the symbols
provided by the linux kernel which runs in the non-root mode will not be
allowed. Spinlock is needed by the early alloc when it is used by
multiple CPUs simultaneously. Although early alloc now is only used
before deprivilege, it will be used by the pKVM hypervisor in the root
mode after the deprivilege. So add pKVM's own spinlock implementation
and use it to protect the memory allocating.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ia86f3d5840f38259a550552ed5eeb727132e157f