ANDROID: KVM: x86/mmu: pKVM: Disable zap_all quirk
When deleting a memslot of a running VM, by default KVM uses
kvm_mmu_zap_all_fast() instead of kvm_mmu_zap_memslot(), i.e. removes
all guest mappings, not just mappings for the pages from this memslot.
That is a workaround for an issue observed in some setups with an Nvidia
GeForce GPU assigned to a qemu VM [1][2]. The root cause of that issue
is still unknown. Quite possibly it cannot be triggered with pKVM, where
guest page table management is fully done in the pKVM hypervisor, with a
completely different code than in normal KVM.

So for simplicity keep this workaround disabled with pKVM, until there
is an evidence that this issue can be triggered in a real pKVM setup.

With this change, kvm_unmap_gfn_range() can remain the single place
where KVM makes the unmap hypercall to pKVM (since kvm_unmap_gfn_range()
is used by both MMU notifiers and memslot deletion).

[1] https://patchwork.kernel.org/project/kvm/patch/20190205210137.1377-11-sean.j.christopherson@intel.com/
[2] https://lore.kernel.org/kvm/20200713190649.GE29725@linux.intel.com/T/

Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: None
Change-Id: Ied3da6939c8f85e698da2422a494c4e15352007e