ANDROID: pKVM: x86: Rearrange the pKVM text/bss/data/rodata sections
To isolate the pKVM hypervisor, eventually pKVM's text/bss/data/rodata
memory should be unmapped from the host EPT so that the deprivileged host
cannot touch any of these. Rearrange these sections to make them in page
aligned memory range so that can be unmapped easily in the future.

The _ASM_EXTABLE() in the pKVM requires to jump to the new rearranged
pKVM text sections, which are required to be authorized. Update the
modpost to include the new rearranged pKVM text sections as part of the
authorized list.

[ Dmytro: Add pKVM text to authorized sections to jump on fault ]

Co-developed-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>
Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I94cf3bd0e94b4a2e43fc662b9fb109266c3f1225