ANDROID: pKVM: VMX: Setup the VMCS before deprivilege
To deprivilege the linux kernel, the VMX should be turned on and VMCS
page should be loaded for each phys CPU. Setup the VMCS control fields
according to pKVM's vmcs_config, which is to minimize the VM exits from
the host by passthrough IO/MSR/EXCEPTIONS. The CR4.VMXE is monitored via
the CR4_GUEST_HOST_MASK, so that changing VMXE bit in CR4 by the host
will only effact the CR4_READ_SHADOW rather than the control register
itself, to prevent the host controling the VMX in real. The VMCS guest
state (for the host) is setup according to the current native state. As
the pKVM hypervisor is not isolated yet, the EPT is not enabled for the
host and the VMCS host state (for the pKVM hypervisor) is still sharing
the same state with the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ib2a6be4ea3cc335ffdedaa8b0d7198a6fab825f1