pKVM: x86: Add host kvm instance
Add a global kvm instance for the host VM. As the kvm instance for Intel
processors is actually pointing to a kvm_vmx structure, reserve the
corresponding memory pages according to the kvm_vmx structure size. As not
all elements in the kvm instance will be useful for the host, the kvm
instance will be initialized on demand.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>