ANDROID: pKVM: VMX: Implement flush_tlb for the host ept
Implement the flush_tlb callback for the host ept to set the request
KVM_REQ_TLB_FLUSH_CURRENT for each host vcpu and kick them out of
non-root mode. And this request is handled by syncing the host ept
context.

As the host ept flushing requires to shoot down all the vcpus, and it
also doesn't have an address granularity flushing, mark the host ept as
flush_tlb_lazy.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Iff1426d018d437f7040c07c239c0ee96fbec0908