ANDROID: pKVM: x86: Move host MMU mapping setup into pkvm_host_mmu_init()
This is to prepare pKVM to use the page protection API to map/unmap
pages for the host mmu. Moving the memory blocks and MMIO mapping into
pkvm_host_mmu_init() can make the pkvm_host_mmu_map() API as a static
function. With this, the only user of pkvm_host_mmu_map() API is to map
host MMIO at the running time, which will be replaced by a dedicated
MMIO map function, or use the MMIO page protection API in the future.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I1edac86f2ebddd777898679eacb8c38745b6248e