ANDROID: pKVM: x86: Redefine kvm_x86_call for pKVM
The original kvm_x86_call relies on the static call mechanism, which is
not supported by the pKVM hypervisor. So redefine as a pKVM-specific
implementation that directly invokes the corresponding operations in
the kvm_x86_ops, or panic if the required operation is missed.

The kvm_pmu_call is intentionally left undefined, as guest PMU support
is not available in pKVM. This also helps catch related compilation
errors early.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ib73d085a4d56e83969a51dd69d43ea3ccf63c9a7