ANDROID: pKVM: x86: Introduce struct pkvm_memcache
Introduce struct pkvm_memcache, which can be used to store a list of
page-aligned physical memory ranges.

The helper push_pkvm_memcache() is to store a memory range's physical
address and size in the pkvm_memcache.

The helper pop_pkvm_memcache() is to retrieve the last stored memory
range's physical address and size from the pkvm_memcache.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ide1f0b8c8cc12148bba1273e5ada8da48517a8e7