x86/pkvm: Use hypercall to start CPUs
pKVM does not allow starting secondary CPUs of protected guests in the
traditional way via INIT/SIPI, since the APIC, including INIT/SIPI, is
emulated by the host and thus is not trusted to control when and at what
address to start running a CPU. Instead, pKVM requires the guest to use
its PKVM_GHC_START_CPU hypercall for that.

Luckily kernel already provides an infrastructure for replacing the
standard method of starting secondary CPUs with a custom one. So
leverage it.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>