ANDROID: pKVM: x86: Add vcpu_put PV interface
Add vcpu_put PV interface for the host to unload a guest vCPU via the
vendor vcpu_put operation. As the current active vCPU is a host vCPU
when involving this PV interface, loading the guest vCPU first before
putting it in case the vendor vcpu_put operation requires. After the
guest vCPU is put, decrements the reference count of the corresponding
pkvm_vcpu instance.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I14ccad082fb40de6cec1f664a3cc2d19ee137777