ANDROID: KVM: pVMX: Implement get_msr operation
Enable the pKVM VMX host to emulate the guest MSRs reading. If a MSR is
allowed to be emulated by the host, then will directly emulate it.
Otherwise use get_msr PV interface to request the pKVM to emulate.
Skipping this PV interface once the guest state is marked as protected
as further reading guest MSRs is no longer permitted by the pKVM
hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ia23ba4c97592000c259d197e0c777437aa0d4914