ANDROID: KVM: pVMX: Implement vcpu_reset operation
Enable the pKVM VMX host to reset npVM's vCPU for init event via the
vcpu_reset PV interface.

Emulating xapic mode will require the host to decode MMIO instructions
which is not supported if the gust is a pVM as pVM's CPU and memory is
isolated from the host. To avoid using xapic mode for a pVM, enabling
x2apic mode by default so that pVM will use MSR instructions to access
lapic, which doesn't require decoding.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I75ddd833317c00a4e1dff597119097db31ea6399