ANDROID: pKVM: x86: Add percpu pkvm_task_struct
Add percpu pkvm_task_struct in pKVM to facilitate the FPU management. It
contains a struct task_struct which is pointed by pKVM's percpu
current_task pointer and a struct fpu which will be used by pKVM to
switch the FPU state between the host VM and pVMs.

The reason of introducing struct task_struct and percpu current_task
pointer is to make the pKVM hypervisor being able to leverage some of
the kernel FPU code which uses current.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I6051d299f77ad55c6e17426678f74f0ca35b56e6