ANDROID: pKVM: x86: Import fpu_swap_kvm_fpstate
Import fpu_swap_kvm_fpstate() from kernel fpu with the modification to
switch the fpstate between host and pVM. Before switching from the host
to a pVM, load the FPU with the pVM's fpstate (the host should save its
own FPU state already before requesting entering a pVM). And before
switching from a pVM to the host, saving the pVM's FPU state and loading
with pKVM's percpu fpu in the current_task, which will configure the FPU
to initial state, so that to protect the pVM's FPU state from the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I728a421cc8c3f7b24acdfc6ef36e6b9196242926