ANDROID: pKVM: x86: Import fpu_enable_guest_xfd_features
Import fpu_enable_guest_xfd_features() from kernel fpu for pKVM to
enable the guest xfd features. The __xfd_enable_feature() is implemented
differently with kernel as pKVM doesn't support allocating memory
dynamically thus the fpstate memory for xfd feature is required to be
allocated and donated from the host via a dedicated PV interface. Thus
the xfd feature enabling in pKVM is to check if the fpstate memory size
is sufficiently large for enabling xfd.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ief5577294611eb9c6612143c31ac6c798f81c6ab