ANDROID: pKVM: VMX: Ensure the host VMCS is loaded when handling NMI
The pKVM captured NMI should be injected to the host VM. With the
vcpu_load PV interface, it is possible that a guest VMCS is currently
loaded when an NMI is captured. To ensure NMI being correctly injected
to the host VM, load the host VMCS before handling NMI and restore the
previous VMCS after that.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Id22a3bb06f0ed74db4bde25add0d7a0e3115cd63