ANDROID: pKVM: VMX: Handle EXIT_REASON_CPUID
As the pKVM hypervisor maintains the guest CPUID entries, the
EXIT_REASON_CPUID is handled alone inside the pKVM hypervisor.

The guest should have a way to detect it is running in a pVM or a npVM
so that it can enable the pVM's enlightenment, emulating the pVM's CPUID
leaf KVM_CPUID_SIGNATURE with the signature "PKVMPKVMPKVM".

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I48e3f82a9e4972dfc17f71c26cd192a821543ce4