ANDROID: pKVM: x86: Add complete_emulated_msr PV interface
Add complete_emulated_msr PV interface for the host to complete an MSR
emulation. For the npVM, the host itself can complete the emulated MSR
by either injecting the exception or skipping the instruction according
to the emulation result. For the pVM, just save the error code rather
than completing the MSR emulation, to prevent the host from injecting
exception or skipping instructions as the host can use this PV interface
at any scenario, e.g, not for MSR emulation at all. The pKVM hypervisor
will decide how to complete the MSR emulation according to the last exit
reason and this saved error code before entering the guest again.

If the host has set an error code through this PV interface, the nmi and
interrupt are not allowed for injection as the pKVM hypervisor will
inject an exception before entering the guest.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I6f3333082264eb4cecee4937d41727af8afc391e