ANDROID: pKVM: VMX: Handle EXIT_REASON_VMCALL
For the npVM, the EXIT_REASON_VMCALL is handled alone by the host. The
exit instruction length is shared with the host to skip the vmcall
instruction.

For the pVM, the EXIT_REASON_VMCALL is handled by the pKVM hypervisor.
Currently no vmcall is supported thus simply return -KVM_EPERM to the
pVM. And #GP will be injected if the vmcall is sent from user mode to
make sure vmcall can only be used in the kernel mode.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Id08d2cba0c3fc3640017dd4941ef1b993d2e2daf