ANDROID: pKVM: VMX: Handle EXIT_REASON_PAUSE_INSTRUCTION
The EXIT_REASON_PAUSE_INSTRUCTION is handled cooperatively by the pKVM
hypervisor and the host. It is handled by the pKVM to grow PLE window
and then returns back to the host to emulate pause. The host will skip
instructions for the npVM while the pKVM will skip for the pVM.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I2f94ab6e571e4954792e2793e35a37693819b7be