ANDROID: pKVM: x86: Handle pVM's IO hypercalls
The pVM's IO hypercalls are used by the pVM to access devices' MMIO
which are emulated by the host. So the hypercalls are emulated by the
pKVM hypervisor to skip the instructions, and eventually handled by the
host via leveraging the kvm_sev_es_mmio_read/write() helpers.

To guarantee the vCPU state protection for the pVM, only necessary vCPU
state are shared with the host (i.e., sharing the RAX/RCX/RDX/RBX/RSI
with the host, and updating the RAX with the result from the host).

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ifa6a4d8e79ab2a29cbf15c2742060ea893ee4451