ANDROID: pKVM: x86: extend kvm_pkvm_vcpu about req_param
Introduce req_param in kvm_pkvm_vcpu which can be used to pass some
extra parameters from pKVM to the host when doing a request, for
example, via pkvm_make_req_to_host().

The motivation for it is the next patch in which pKVM requests memcache
refill from the host by forwarding the PKVM_GHC_SHARE_MEM hypercall to
the host (which happens via pkvm_make_req_to_host()) and there is a need
to inform the host about refill size.

An alternative solution could be to pass it in a hypercall arg register
not used by the given hypercall issued by the guest, e.g. in rsi in this
case of PKVM_GHC_SHARE_MEM, but that would be unsafe because if the
guest issues the hypercall via inline asm, e.g. the following one in
Linux guests:

	asm volatile(KVM_HYPERCALL
		     : "=a"(ret)
		     : "a"(nr), "b"(p1), "c"(p2), "d"(p3), "S"(p4)
		     : "memory");

the compiler expects that only rax ("=a" as an output operand) will be
modified.  Modifying other registers not being part of output operand or
not being part of clobber list is wrong as the compiler doesn't expect
modification of those registers and is free to generate code with such
assumption.

[ Dmytro: s/int/u64/ ]

Signed-off-by: Grzegorz Jaszczyk <jaszczyk@google.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: None
Change-Id: Ida2c426e480dec9326e08ba9e76914fe50042fcd