ANDROID: pKVM: VMX: Handle EXIT_REASON_MSR_READ/WRITE_IMM
The EXIT_REASON_MSR_READ/WRITE_IMM exits are handled similar with the
traditional MSR exits, which are handled by the pKVM hypervisor first.
If cannot be handled but allowed to be handled by the host, returns back
to the host from the vcpu_run PV interfaces, meanwhile with sharing the
necessary vCPU state with the host, e.g., exit instruction information
and instruction length, and the corresponding registers saving the
immediate value. After this is handled by the host, the pKVM will update
the corresponding immediate register with the result from the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I91c5ee013fbe9d3cff349be031f6072678b2ce61