ANDROID: KVM: pVMX: Implement vm_init operation
Enable the pKVM VMX host to initialize a guest VM using the vm_init PV
interface. The host allocates memory pages to serve as the backing for
the pkvm_vm structure, which is private to the pKVM hypervisor, and gets
the guest VM handle from the vm_init PV interface.

As the pKVM hypervisor can only handle with the contiguous physical
memory, use kzalloc rather than __vmalloc to allocate the kvm structure
in the host.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I44bfb0125270af7a39d50d8404915f13cd69c8ff