ANDROID: KVM: pVMX: Implement hardware_setup/unsetup operations
The host still needs to access the VMX MSRs to know if a feature is
valid or not. And the host also relies on the kvm_caps and kvm_cpu_caps
to do proper initialization, e.g, the initial CPUID for the guest. All
of these are done by the vmx_hardware_setup(). So reuse the
vmx_hardware_setup/unsetup() for the hardware_setup/unsetup operations.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I9bbafea5434a6d8109c3dfc0adc3be76925f270d