ANDROID: pKVM: VMX: Implement vcpu_put operation
Import the vmx_vcpu_put() for the pKVM to implement vcpu_put operation.
Unlike KVM, pKVM lacks an SMP call mechanism to notify remote CPUs to
clear VMCS if they have loaded this vCPU. Therefore, VMCS clearing is
performed immediately when doing the vcpu_put.

Note: The vcpu_put operation does not handle posted interrupts, as
interrupt management is delegated to the host. Additionally, preparing to
switch to host is required when the vCPU exits its run loop but not here
when putting a vCPU.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I78ab4ca2a59bdbbcd10c02d0dc126877f90e1969