pKVM: VMX: Introduce prepare_switch_to_host
Import vmx_prepare_switch_to_host from KVM and use it as the new
operation prepare_switch_to_host in the kvm_x86_ops for the pKVM
hypervisor, to pair with the existing prepare_switch_to_guest operation.
The pKVM hypervisor's segments (only cares about the GDT/IDT/FS/GS) are
restored by the VMX automatically when exiting from the guest. And
restore the MSR_KERNEL_GS_BASE with the host VM's value. This can hide
the guest's value from the host which can provide security benefit for
pVMs. With this, the host doesn't need to save/restore this MSR by
itself.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>