ANDROID: pKVM: x86: trace: Add tracing of host hypercalls
Currently pKVM's vmexit tracing provides perf stats for individual
vmexit reasons only, so in particular, it doesn't distinguish between
different hypercalls (vmexits with VMCALL reason). Whereas in practice
the vast majority of host vmexits to pKVM are hypercalls. So it would be
very useful to provide separate stats for different types of hypercalls
(i.e. based on the hypercall number in the RAX register). So provide it,
in addition to the existing vmexit stats.

No such support for guest hypercalls yet.

[ Chuanxiao: Refactor print_perf_data() a bit ]

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>
Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: None
Change-Id: Id0a5eb13af6d2b2c04fdaeeb0f744d523e415e96