ANDROID: pKVM: x86: Add vcpu_free PV interface
Add vcpu_free PV interface for host to release a vCPU from a guest VM.
This interface cleans up the operations done when creating a vCPU, which
includes detaching the pkvm_vcpu structure, cleaning up vendor
resources, unsharing the host's kvm_vcpu structure and tearing down the
donated memory pages.

The physical address and size of the torn-down memory pages are stored
in a pkvm_memcache structure and returned for the host to reclaim.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ia0dcf50f19552650bd4c907d55715de73c525489