ANDROID: pKVM: x86: Enforce initial values of most MSRs
Even though most MSRs are emulated by pKVM itself, not by the host,
pKVM still allows the host to set initial values of any MSRs for pVMs,
which is a potential source of security issues.

So implement a simple policy similar to the one we implemented for
initial CPUID values: allow the host to pre-initialize only those
MSRs that are currently pre-initialized by crosvm.

The crosvm may still initialize some MSRs which are protected by the
pKVM hypervisor(e.g, MSR_STAR). Such writing will be ignored and a
success result will be returned to the host, to make the crosvm happy
and meanwhile guarantee these MSRs won't be manipulated by the host.
Eventually the crosvm will be fixed by not initializing these MSRs.

MSRs that are managed by the host (e.g. APIC, TSC) are not affected,
since the host doesn't need to ask pKVM to set them anyway.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: None
Change-Id: Ifc21586e933ced80e36f996ae0616df68e7b9800