ANDROID: pKVM: x86: Don't map page 0 in the hypervisor address space
Currently pKVM's vmemmap is mapped starting from virtual address 0 in
the hypervisor page tables. As a result, NULL pointer dereference bugs
in pKVM code go unnoticed, as they result in pKVM reading the 1st page
of the vmemmap (or worse, writing to it), instead of immediate crashing.

So shift the vmemmap's mapping by one page, to keep page 0 unmapped,
so that NULL dereferences in pKVM code result in pKVM crashing with a
page fault.

Signed-off-by: Dmytro Maluka <dmaluka@google.com>

Bug: None
Change-Id: I3b53e641b542bf4a6baadb399365b2df11132409