pKVM: x86: Put the exception table section to rodata section
The pKVM's __ex_table section contains the data which will be used by
the exception handler to fix the exceptions. It may be modified at the
init time due to sort but will not be changed at the run time. Put it
into pKVM's rodata section, which is read-only and isolated from the
host.

The objtool is also updated to handle the pKVM's __ex_table section.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>