ANDROID: pKVM: x86: Setup the VMCS host state fields with pKVM's own state
To isolate the pKVM hypervisor and the deprivileged host, the pkvm
hypervisor should use its own CPU state rather than the host's. Setup
the VMCS host state with pKVM's own GDT/IDT/TSS and so on.

As the pKVM hypervisor runs with interrupt disabled, and no exception
should be triggered if everything is fine, the FRED feature is disabled
as it doesn't have much benefit to the pKVM hypervisor's event delivery.
In case any exception happened, it is handled by a default handler via
executing hlt instruction to stop.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I68800aadfd84fd0131d55a379d8a961e27504df3