pKVM: x86: Add set_efer PV interface
Add set_efer PV interface to allow the host to set the guest EFER MSR
via the vendor-specific set_efer operation. For a pVM, this PV interface
allows the host to set the initial value of the EFER MSR before the pVM
has started running, but doesn't allow to modify it if the pVM is
already running.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>