ANDROID: pKVM: x86: Extend pkvm_fixup_exception to support EX_TYPE_DEFAULT
Extend the pkvm_fixup_exception() to support the EX_TYPE_DEFAULT type.
It is handled by the default rule which is to update the ip to the
desired destination address, as well as the corresponding registers.

The motivation is to support exception handling in the KVM's VMX
instruction helpers, as well as any other instruction which uses the
_ASM_EXTABLE for the exception handling.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I53577fb66f97a4579cbdc06d5634f1c0882171da