ANDROID: pKVM: x86: Audit PV interfaces for protected VMs
As the pVM's state should be isolated by the pKVM hypervisor from the
host, the pKVM hypervisor should check if a PV interface is allowed for
the host to use or not.

The current policy is to simply allow all PV interfaces for an npVM as
there is no isolation between an npVM and the host, but dis-allow all PV
interfaces for a pVMs by default.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I94d67f095161838053fec5cf5486b228e9ff88ca