pKVM: x86: Add set_cr4 PV interface
Add set_cr4 PV interface to allow the host to set guest CR4 via the
vendor set_cr4 operation. For a pVM, this PV interface allows the host
to set the CR4 before the pVM has started running, but doesn't allow to
modify it if the pVM is already running.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>