ANDROID: pKVM: x86: Add set_cr0 PV interface
Add set_cr0 PV interface to allow the host to set guest CR0 via the
vendor-specific set_cr0 operation. For a pVM, this PV interface allows
the host to set the CR0 before the pVM has started running, but doesn't
allow to modify it if the pVM is already running.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: If327a224801791fca9b7008d91e6a73a1f075727