ANDROID: pKVM: x86: Add idt/gdt PV interfaces
Add idt/gdt PV interfaces for the host to access guest IDT and GDT via
the vendor-specific idt/gdt operations. For a pVM, these PV interfaces
allows the host to access the pVM's IDT and GDT before the pVM has
started running, but doesn't allow to access these if the pVM is already
running.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I0b1f119f5b24763ec88683d879c1af3b713365ab