pKVM: x86: Add vcpu_after_set_cpuid PV interface
Add vcpu_after_set_cpuid PV interface for the host to send the CPUID
entries to the pKVM hypervisor, which will be used as the CPUID entries
for the pkvm_vcpu. The CPUID entries memory pages will be donated to
prevent the host from accessing. If the pkvm_vcpu already has a CPUID
entries, the old one's physical addresses and size will be return to the
host via pkvm_memcache.

For a pVM, this PV interface allows the host to set CPUID entries before
the pVM has started running, but doesn't allow the host to set if the
pVM is already running.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>