ANDROID: pKVM: x86: Add vcpu_add_fpstate PV interface
Add vcpu_add_fpstate PV interface for the host to add a new fpstate
memory if needs to support xfd, as the fpstate memory allocated when
creating vcpu may not be sufficient. As the pVM's FPU state is managed
by the pKVM hypervisor while the npVM's FPU state is managed by the
host, re-allocating the fpstate is only necessary for the pVM. This PV
interface is not accessible to the host once the pVM started running to
ensure the host won't manipulate pVM's FPU. The old fpstate memory
physical addresses and size are returned via pkvm_memcache for the host
to reclaim.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I06dfac61964f206907c172f6cf049612d8d4aac7