pKVM: x86: Add irq/nmi mask PV interfaces
Add irq/nmi mask PV interfaces for the host to access irq/nmi mask via
vendor-specific operations. Specifically, __pkvm__interrupt/nmi_allowed
are for the host to check if irq/nmi is allowed or not, for injection or
non-injection, and __pkvm__get/set_nmi_mask are for the host to control
the guest's NMI mask.

As pVM's interrupt are virtualized by the host, the host still needs to
check for irq/nmi allowance or nmi mask for injection. The PV interfaces
__pkvm__interrupt/nmi_allowed, and __pkvm__get_nmi_mask are always
allowed for the host.

As set_nmi_mask is used to emulate, i.e., SMM mode, real mode iret and
setting segment by the host. None of these cases are supported for the
pVM. So the __pkvm__set_nmi_mask is always inaccessible to the host to
avoid the host changing pVM's NMI mask without a real need.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>