ANDROID: pKVM: x86: Add write_tsc_offset/multiplier PV interfaces
Add write_tsc_offset/multiplier PV interfaces to allow host to write the
tsc offset and tsc multiplier. These PV interfaces are always accessible
to the host as the Trusty pVM can use GSC as the secure time source not
the TSC. The same reason to allow host to emulate and access the guest
TSC_ADJUST and TSC MSRs.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: I0dcdcc4878b008d193f571c2d2d8a2b9e14624a1