ANDROID: pKVM: x86: Add KVM_X86_SW_PROTECTED_VM as supported vm_types in pKVM
KVM_X86_SW_PROTECTED_VM is for development and testing purpose. It won't
break the related path if pKVM is enabled. The host side reports this
vm_type is supported, also add it to the pKVM side to make it consistent.

This also fix some kselftest failures, e.g., guest_memfd_test.

Signed-off-by: Chenyi Qiang <chenyi.qiang@intel.com>

Bug: None
Change-Id: I8d8ca97a72b7f09036b911f68ed52e225772a9cf