ANDROID: KVM: pVMX: Implement set_msr operation
Enable the pKVM VMX host to emulate the guest MSRs writing. If a MSR is
allowed to be emulated by the host, then will directly emulate it.
Otherwise use set_msr PV interface to request the pKVM to emulate.
Skipping this PV interface once the guest state is marked as protected
as further setting guest MSRs is no longer permitted by the pKVM
hypervisor.

Signed-off-by: Chuanxiao Dong <chuanxiao.dong@intel.com>

Bug: None
Change-Id: Ied48019bde698eabedb64d713f8a3f6235004104